requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
    #   starlette
asttokens==3.0.0
    # via stack-data
asyncpg==0.30.0
    # via jouster (pyproject.toml)
black==25.1.0
    # via jouster (pyproject.toml)
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   black
    #   nltk
    #   uvicorn
decorator==5.2.1
    # via ipython
distro==1.9.0
    # via openai
executing==2.2.1
    # via stack-data
fastapi==0.116.1
    # via jouster (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
ipython==9.5.0
    # via jouster (pyproject.toml)
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jiter==0.10.0
    # via openai
joblib==1.5.2
    # via nltk
jsonpatch==1.33
    # via langchain-core
jsonpath-ng==1.7.0
    # via redisvl
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via jouster (pyproject.toml)
langchain-core==0.3.75
    # via
    #   jouster (pyproject.toml)
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==0.3.32
    # via jouster (pyproject.toml)
langchain-text-splitters==0.3.11
    # via langchain
langgraph==0.6.6
    # via jouster (pyproject.toml)
langgraph-checkpoint==2.1.1
    # via
    #   langgraph
    #   langgraph-checkpoint-postgres
    #   langgraph-checkpoint-redis
    #   langgraph-prebuilt
langgraph-checkpoint-postgres==2.0.23
    # via jouster (pyproject.toml)
langgraph-checkpoint-redis==0.1.1
    # via jouster (pyproject.toml)
langgraph-prebuilt==0.6.4
    # via langgraph
langgraph-sdk==0.2.5
    # via langgraph
langsmith==0.4.23
    # via
    #   jouster (pyproject.toml)
    #   langchain
    #   langchain-core
matplotlib-inline==0.1.7
    # via ipython
ml-dtypes==0.5.3
    # via redisvl
mypy-extensions==1.1.0
    # via black
nltk==3.9.1
    # via jouster (pyproject.toml)
numpy==2.3.2
    # via
    #   ml-dtypes
    #   redisvl
openai==1.105.0
    # via langchain-openai
orjson==3.11.3
    # via
    #   langgraph-checkpoint-postgres
    #   langgraph-checkpoint-redis
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   black
    #   langchain-core
    #   langsmith
    #   pytest
parso==0.8.5
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
platformdirs==4.4.0
    # via black
pluggy==1.6.0
    # via pytest
ply==3.11
    # via jsonpath-ng
prompt-toolkit==3.0.52
    # via ipython
psycopg==3.2.9
    # via langgraph-checkpoint-postgres
psycopg-pool==3.2.6
    # via langgraph-checkpoint-postgres
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.11.7
    # via
    #   jouster (pyproject.toml)
    #   fastapi
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
    #   redisvl
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
pytest==8.4.1
    # via jouster (pyproject.toml)
python-dotenv==1.1.1
    # via jouster (pyproject.toml)
python-ulid==3.1.0
    # via redisvl
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
    #   redisvl
redis==6.4.0
    # via
    #   langgraph-checkpoint-redis
    #   redisvl
redisvl==0.8.2
    # via langgraph-checkpoint-redis
regex==2025.9.1
    # via
    #   nltk
    #   tiktoken
requests==2.32.5
    # via
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.43
    # via langchain
stack-data==0.6.3
    # via ipython
starlette==0.47.3
    # via fastapi
tenacity==9.1.2
    # via
    #   langchain-core
    #   redisvl
tiktoken==0.11.0
    # via langchain-openai
tqdm==4.67.1
    # via
    #   nltk
    #   openai
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   ipython
    #   langchain-core
    #   openai
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via jouster (pyproject.toml)
wcwidth==0.2.13
    # via prompt-toolkit
xxhash==3.5.0
    # via langgraph
zstandard==0.24.0
    # via langsmith
